requirements.txt
telethon==1.34.0
playwright==1.40.0
google-generativeai==0.3.2
asyncio
tzdata
